requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.13.4
blis==0.7.11
catalogue==2.0.10
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.0
coqpit-config==0.1.1
coqui-tts==0.25.3
coqui-tts-trainer==0.2.0
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
dateparser==1.1.8
decorator==5.2.1
docopt==0.6.2
EbookLib==0.19
einops==0.8.1
encodec==0.1.1
ffmpeg==1.4
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2024.6.1
grpcio==1.73.1
gruut==2.4.0
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
hf-xet==1.1.5
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==7.5.0
Jinja2==3.1.3
joblib==1.5.1
jsonlines==1.2.0
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
lxml==6.0.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
monotonic-alignment-search==0.2.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
murmurhash==1.0.13
networkx==3.2.1
num2words==0.5.14
numba==0.60.0
numpy==1.26.3
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
onnxruntime==1.19.2
packaging==25.0
pillow==11.0.0
piper-phonemize==1.1.0
piper-tts==1.2.0
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.10
progressbar==2.5
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
PyPDF2==3.0.1
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rotary-embedding-torch==0.8.8
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
SudachiDict-core==20250515
SudachiPy==0.6.10
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.6.0
tokenizers==0.20.3
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.67.1
transformers==4.46.2
triton==3.3.1
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
Unidecode==1.4.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
